Refactor AST type system to enhance type inference and alias handling; integrate SymbolTable in type comparisons, simplify recursive type resolution, and consolidate type validation logic

+++ packages/compiler/include/ast/definitions/function_definition.h
@@ -90,7 +90,7 @@ namespace stride::ast::definition
 
         bool matches_type_signature(const std::string& name, const AstFunctionType* signature) const;
 
-        bool matches_generic_signature(const std::string& name, int function_param_count, int generic_param_count) const;
+        bool matches_generic_signature(const std::string& name, size_t function_param_count, size_t generic_param_count) const;
 
         void set_llvm_function(llvm::Function* function)
         {
